script-generation/pyproject.toml
[project]
name = "script-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=23.0.0",
    "colorama>=0.4.6",
    "flake8>=6.0.0",
    "mistralai>=1.9.11",
    "openai>=1.0.0",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]
